refactor: restructure of all utility/config/tooling files
