+++ .prettierignore
@@ -1,1 +1,1 @@
-utils/linting/.prettierignore
+linting/.prettierignore
