+++ package.json
@@ -5,22 +5,22 @@
   "description": "Strimzi UI provides a way for managing Strimzi and Kafka clusters (+ other components) deployed by it using a graphical user interface.",
   "scripts": {
     "clean": "rm -rf dist/ && rm -rf generated/",
-    "start": "webpack-dev-server --config utils/build/webpack.dev.js",
-    "build": "webpack --config utils/build/webpack.prod.js",
+    "start": "webpack-dev-server --config build/webpack.dev.js",
+    "build": "webpack --config build/webpack.prod.js",
     "server": "",
-    "test:unit": "jest --collectCoverage --config test/jest.config.js",
+    "test:unit": "jest --collectCoverage --config test_common/jest.config.js",
+    "test:e2e": "npm-run-all clean build server cy:closed",
     "test:e2e:open": "npm-run-all clean build server cy:open",
-    "test:e2e:closed": "npm-run-all clean build server cy:closed",
-    "pretest": "npm-run-all --parallel licence:check && prettier --config ./utils/linting/prettier.config.js --check .",
+    "pretest": "npm-run-all --parallel licence:check && prettier --config ./linting/prettier.config.js --check .",
     "storybook": "start-storybook --port 6006",
     "storybook:build": "build-storybook -o .out",
-    "licence:check": "license-check-and-add --f ./utils/linting/license-check-and-add.config.json check",
-    "licence:add": "license-check-and-add --f ./utils/linting/license-check-and-add.config.json add",
+    "licence:check": "license-check-and-add --f ./linting/license-check-and-add.config.json check",
+    "licence:add": "license-check-and-add --f ./linting/license-check-and-add.config.json add",
     "lint": "npm-run-all 'lint:allfiles -- .' 'lint:js -- .' 'lint:styling -- \"**/*.(s)css\"' 'lint:format -- .'",
     "lint:allfiles": "npm-run-all 'licence:add -- {1}' --",
-    "lint:js": "eslint --config ./utils/linting/eslint.config.js --fix --max-warnings 0 --ignore-path ./utils/linting/.eslintignore",
-    "lint:styling": "stylelint --config ./utils/linting/stylelint.config.js --fix --max-warnings 0",
-    "lint:format": "prettier --config ./utils/linting/prettier.config.js --write",
+    "lint:js": "eslint --config ./linting/eslint.config.js --fix --max-warnings 0 --ignore-path ./linting/.eslintignore",
+    "lint:styling": "stylelint --config ./linting/stylelint.config.js --fix --max-warnings 0",
+    "lint:format": "prettier --config ./linting/prettier.config.js --write",
     "cy:open": "cypress open --config-file e2e/cypress.json",
     "cy:closed": "cypress run --config-file e2e/cypress.json"
   },
@@ -52,6 +52,7 @@
     "@storybook/addon-links": "^6.0.22",
     "@storybook/addon-storysource": "^6.0.22",
     "@storybook/react": "^6.0.22",
+    "@types/cypress-cucumber-preprocessor": "^1.14.1",
     "@types/jest": "^26.0.14",
     "@types/node": "^14.11.2",
     "babel-eslint": "^10.1.0",
