+++ test_common/jest.config.js
@@ -2,9 +2,10 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-const { jestModuleMapper } = require('../aliasHelper.js');
+const { jestModuleMapper } = require('../utils/tooling/aliasHelper.js');
 
 const config = {
+  rootDir: '../',
   clearMocks: true,
   coverageDirectory: 'coverage',
   coveragePathIgnorePatterns: [
@@ -12,7 +13,11 @@ const config = {
     '/dist/',
     '/e2e/',
     '/utils/',
+    '/linting',
+    '/build/',
     '/coverage/',
+    '/.storybook/',
+    '/test_common/',
   ],
   collectCoverageFrom: ['**/*.{js,jsx,ts,tsx}', '!**/*.config.*'],
   coverageReporters: ['json', 'text', 'lcov', 'json-summary'],
@@ -31,12 +36,15 @@ const config = {
   testPathIgnorePatterns: ['/e2e/'],
   transform: {
     '^.+\\.(ts|tsx)$': 'ts-jest',
-    '^.+\\.(js|jsx|ts|tsx)$': ['babel-jest', { configFile: './test.babel.js' }],
+    '^.+\\.(js|jsx|ts|tsx)$': [
+      'babel-jest',
+      { configFile: './test_common/test.babel.js' },
+    ],
   },
   moduleNameMapper: {
     ...jestModuleMapper,
   },
-  setupFiles: ['./support'],
+  setupFiles: ['./test_common/jest_cucumber_support'],
 };
 
 module.exports = config;
