+++ test_common/test.babel.js
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-const babelPresets = require('../utils/build/babelPresets.js');
+const babelPresets = require('../build/babelPresets.js');
 
 module.exports = {
   presets: babelPresets,
