+++ tsconfig.json
@@ -3,7 +3,7 @@
     "strict": true,
     "target": "es5",
     "lib": ["es5", "dom"],
-    "types": ["node"]
+    "types": ["node", "jest"]
   },
   "include": ["**/*.ts"]
 }
