--- utils/linting/README.md
@@ -1,3 +1,0 @@
-# Linting
-
-This directory contains code and configuration required to lint this codebase, as per [the linting documentation](../../docs/Linting.md). Some of the tools used require configuration to exist at the root of the project. In these cases, symlinks are used to reference the actual configuration in this directory. All files in this directory are named to indicate what tool they are used with, and will contain comments to reference how they implement/enable the configuration desired.
