+++ utils/tooling/README.md
@@ -1,0 +1,7 @@
+# Tooling
+
+This directory contains any code that is used for development coding. For example, code that aids linting or the Webpack build process.
+
+## Contents
+
+- [headers](./headers/README.md) - header to be inserted at the top of every file.
