+++ utils/tooling/aliasHelper.js
@@ -6,7 +6,8 @@
 const path = require('path');
 const fs = require('fs');
 
-const src = path.resolve(__dirname, '../client');
+const client = '../../client';
+const src = path.resolve(__dirname, client);
 const moduleNames = fs.readdirSync(src);
 
 // Aliases relative to the current file
@@ -18,13 +19,14 @@ const relativeAliases = {
 
 const relativeMocks = Object.keys(relativeAliases).join('|');
 const ignoredBinaries = ['png', 'svg', 'ico', 'scss'].join('|');
+const mockFile = '<rootDir>/test_common/mockfile.util.js';
 
 const modules = moduleNames.reduce((moduleConfig, name) => {
   const currentModule = {
-    path: path.resolve(__dirname, `../client/${name}`),
+    path: path.resolve(__dirname, `${client}/${name}`),
     mapper: {
       regex: `^${name}(.*)$`,
-      path: `<rootDir>/client/${name}$1`,
+      path: `${client}/${name}$1`,
     },
   };
   return { ...moduleConfig, [name]: currentModule };
@@ -46,8 +48,8 @@ const jestModuleMapper = Object.values(modules).reduce(
     return { ...mapping, [regex]: path };
   },
   {
-    [`^(${relativeMocks})(.*)$`]: '<rootDir>/test/mockfile.util.js',
-    [`^.+\\.(${ignoredBinaries})$`]: '<rootDir/test/mockfile.util.js',
+    [`^(${relativeMocks})(.*)$`]: mockFile,
+    [`^.+\\.(${ignoredBinaries})$`]: mockFile,
   }
 );
